+++ changelog.md
@@ -1,3 +1,12 @@
 # Changelog
 
-Snapshot 2021-12-04 15:48
+**./test-parts/possible/valid/template-criteria.js**
+* Merged common test code into new function 'handlePathfinding'
+	* Perform pathfinding.
+	* Check base results.
+	* Check routes against template.
+	* Return number of routes.
+* Changes to the "Exact" test:
+	* Declared 'resultCount' variable.
+	* 'resultCount' is assigned with 'handlePathfinding'
+	* Replaced 'resultValue.length' reference with 'resultCount'